requirements.txt
langchain==0.0.250
nltk==3.8.1
numpy==1.24.3
openai==0.27.8
pandas==2.0.1
python-dotenv==1.0.0
pymongo==4.4.0
regex==2023.6.3
requests==2.30.0
retry==0.9.2
tiktoken==0.4.0
streamlit==1.26.0
watchdog==3.0.0